requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
anthropic>=0.34.0
reportlab==4.2.2
python-barcode[images]==0.15.1
Pillow==10.4.0
